daily-brief/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
feedparser>=6.0.11
jinja2>=3.1.4
twilio>=9.4.0
pymongo>=4.10.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
